+++ frontend/scripts/react-components/tool/mapbox-map/layers/base-layer.js
@@ -1,0 +1,10 @@
+export const baseLayer = (baseLayerInfo) => ({
+  id: baseLayerInfo.id,
+  type: 'raster',
+  source: {
+    type: 'raster',
+    tiles: [baseLayerInfo.url],
+    minzoom: 2,
+    maxzoom: 12
+  }
+});
